requirements.txt
flask
gunicorn
Pillow